Fix for Sequel 3.34.1
Change in 3.34:
"Database#uri and url now return the connection string given to
Sequel.connect. Previously, they tried to reconstruct the url using the
database's options, but that didn't work well in corner cases."

Lo and behold there's a database_type() method anyway and it was there
in 3.33 all along.