Use a single statement for DataMapper truncation on PostgreSQL.
This change aligns the behavior of the DM PostgreSQL truncation adapter with
the ActiveRecord one. A single statement is faster and safer than lots of
`TRUNCATE TABLE ... CASCADE` statements.

While this commit only implements single statements for PostgreSQL, it
provides the extension point necessary for other adapters to switch to single
statements in the future.