Memoize tables for truncation
As explained in issue #130 active record connections don't do caching of
schema information. This meant that database_cleaner did a query every
with the truncation strategy for every clean to find out the tables. Now
we cache the information. If someone needs cache invalidation then it
can be added. As we already did view caching this is consistent between
views and tables.