Fix an issue when the mysql2 gem is included in a project that uses non-msql databases.
The old implementation assumed if the gem was present, that all
databases were mysql. Connecting to a DB of any other type would cause
exceptions. Instead, we check the connection adapter type in each
instance of the cleaner, and only use the mysql2 specific code when
using the mysql2 adapter