Fix truncation method for old versions of mongodb
the mongo gem removed the `Mongo::VERSION` constant for versions 1.8.x to 1.12.x till it returned in version 2 and above. for super old versions of mongo, namely 1.7.x and below the constant still existed, so we need to check definition of the version constant and check the version number to correctly select the truncation method name.