Add DatabaseCleaner::strategy getter method
Issue #511 proposes the addition of a ::strategy getter
on the DatabaseCleaner module to aid new users with debugging.
This commit introduces that method.

Like the ::strategy= setter method, ::strategy is actually defined
as an instance method on DatabaseCleaner::Cleaners,
and is made available to the top-level module by delegation.

Depending on the complexity of the user's configuration,
this method may return three different kinds of values:

1. nil (if no strategies have been set on any cleaners);
2. a single strategy (if all cleaners share the same strategy); or
3. a hash of cleaner IDs and the strategies that go with them
   (if multiple strategies have been set on multiple cleaners).

The first of these three cases has not explicitly been tested.

=== Note

This commit also restores a spec context that was removed in c542fee:

    DatabaseCleaner::Cleaners
      top level api methods
        multiple cleaners
          multiple orm proxy methods
            with differing orms and dbs

This spec context included a single example testing the #orm= method,
which was removed in the aforementioned commit.
Now, a modified version of the same context is used
to test the #strategy method introduced here.